+++ sdk/ai/azure-ai-agents/CHANGELOG.md
@@ -8,10 +8,13 @@
 ### Features Added
 
 - Add `RunStepDetailsActivity`, describing MCP function parameters.
+- Add `RunStepDeltaCustomBingGroundingToolCall`, describing `BingCustomSearchTool` updates in streaming scenario.
 
 ### Bugs Fixed
 
 ### Sample updates
+
+- Bing Grounding and Bing Custom Search samples were fixed to correctly present references.
 
 ## 1.2.0b2 (2025-08-12)